+++ src/dr_clients/include/stack.h
@@ -12,7 +12,7 @@ typedef struct _stack_t {
 } stack_t;
 
 
-bool stack_init(stack_t ** stack, uint capacity, void(*free_data_func) (void *));
+bool stack_init(stack_t ** stack, uint64 capacity, void(*free_data_func) (void *));
 void stack_delete(stack_t * stack);
 void * stack_pop(stack_t * stack);
 void * stack_peek(stack_t * stack);
